refactor for multiple tracks
